src/requirements.txt
cassandra-driver==3.7.0
futures==3.0.5
googlefinance==0.7
kafka-python==1.3.1
schedule==0.3.2
six==1.10.0
